Improve based on comments
