+++ packages/flutter/test/widgets/tween_animation_builder_test.dart
@@ -393,169 +393,4 @@ void main() {
     await tester.pump(const Duration(seconds: 2));
     expect(values, <Size>[const Size(10, 10)]);
   });
-
-  group('RepeatingTweenAnimationBuilder', () {
-    testWidgets('Repeats animation continuously', (WidgetTester tester) async {
-      final List<double> values = <double>[];
-      const Duration duration = Duration(milliseconds: 100);
-
-      await tester.pumpWidget(
-        RepeatingTweenAnimationBuilder<double>(
-          duration: duration,
-          animatable: Tween<double>(begin: 0.0, end: 1.0),
-          builder: (BuildContext context, double value, Widget? child) {
-            values.add(value);
-            return const Placeholder();
-          },
-        ),
-      );
-
-      // Initial value should be 0.0.
-      expect(values.last, 0.0);
-
-      // Pump to a frame just before the end of the first cycle.
-      await tester.pump(const Duration(milliseconds: 99));
-      // The value should be very close to 1.0.
-      expect(values.last, moreOrLessEquals(0.99));
-
-      // Pump past the cycle boundary (1ms to complete the cycle + 50ms into the next cycle).
-      await tester.pump(const Duration(milliseconds: 51));
-      // The animation should have wrapped around and be at the halfway point of the next cycle.
-      expect(values.last, moreOrLessEquals(0.5));
-    });
-
-    testWidgets('Reverses animation when reverse is true', (WidgetTester tester) async {
-      final List<double> values = <double>[];
-
-      await tester.pumpWidget(
-        RepeatingTweenAnimationBuilder<double>(
-          duration: const Duration(milliseconds: 100),
-          animatable: Tween<double>(begin: 0, end: 1),
-          reverse: true,
-          builder: (BuildContext context, double value, Widget? child) {
-            values.add(value);
-            return const Placeholder();
-          },
-        ),
-      );
-
-      // Let it run long enough to reverse.
-      for (int i = 0; i < 20; i++) {
-        await tester.pump(const Duration(milliseconds: 20));
-      }
-
-      // Should have both increasing and decreasing values.
-      expect(values.first, 0.0);
-      expect(values.any((double v) => v == 1.0), isTrue, reason: 'Should reach max');
-
-      // Check for reversing pattern - values should decrease after reaching 1.
-      bool foundReverse = false;
-      for (int i = 1; i < values.length; i++) {
-        if (values[i] < values[i - 1] && values[i - 1] > 0.5 && values[i] < 0.9) {
-          foundReverse = true;
-          break;
-        }
-      }
-      expect(foundReverse, isTrue, reason: 'Should have reversing motion');
-    });
-
-    testWidgets('Handles pause and unpause correctly', (WidgetTester tester) async {
-      final List<int> values = <int>[];
-      Widget buildWidget({required bool paused}) {
-        return RepeatingTweenAnimationBuilder<int>(
-          duration: const Duration(seconds: 1),
-          animatable: IntTween(begin: 0, end: 100),
-          paused: paused,
-          builder: (BuildContext context, int value, Widget? child) {
-            values.add(value);
-            return const Placeholder();
-          },
-        );
-      }
-
-      // Start animated.
-      await tester.pumpWidget(buildWidget(paused: false));
-      await tester.pump(const Duration(milliseconds: 500));
-      expect(values, <int>[0, 50]);
-      // Pause mid-animation.
-      await tester.pumpWidget(buildWidget(paused: true));
-      final int pausedValue = values.last;
-      await tester.pump(const Duration(seconds: 2));
-      expect(values.last, pausedValue); // Should stay paused.
-      // Resume animation.
-      await tester.pumpWidget(buildWidget(paused: false));
-      await tester.pump(const Duration(milliseconds: 400));
-      expect(values.last, greaterThan(pausedValue)); // Should resume from where it paused.
-    });
-
-    testWidgets('Does not animate when begin equals end', (WidgetTester tester) async {
-      final List<int> values = <int>[];
-      await tester.pumpWidget(
-        RepeatingTweenAnimationBuilder<int>(
-          duration: const Duration(seconds: 1),
-          animatable: IntTween(begin: 100, end: 100),
-          builder: (BuildContext context, int value, Widget? child) {
-            values.add(value);
-            return const Placeholder();
-          },
-        ),
-      );
-      expect(values, <int>[100]);
-
-      await tester.pump(const Duration(seconds: 2));
-      expect(values, <int>[100]); // Should not repeat when there's no animation.
-    });
-
-    testWidgets('Passes child to builder correctly', (WidgetTester tester) async {
-      const Widget childWidget = Text('Child');
-      Widget? receivedChild;
-
-      await tester.pumpWidget(
-        Directionality(
-          textDirection: TextDirection.ltr,
-          child: RepeatingTweenAnimationBuilder<double>(
-            duration: const Duration(seconds: 1),
-            animatable: Tween<double>(begin: 0.0, end: 1.0),
-            builder: (BuildContext context, double value, Widget? child) {
-              receivedChild = child;
-              return child ?? const Placeholder();
-            },
-            child: childWidget,
-          ),
-        ),
-      );
-
-      expect(receivedChild, childWidget);
-      expect(find.text('Child'), findsOneWidget);
-    });
-
-    testWidgets('Supports animatables without explicit begin/end', (
-      WidgetTester tester,
-    ) async {
-      final List<double> values = <double>[];
-
-      await tester.pumpWidget(
-        RepeatingTweenAnimationBuilder<double>(
-          duration: const Duration(milliseconds: 100),
-          animatable: CurveTween(curve: Curves.easeIn),
-          builder: (BuildContext context, double value, Widget? child) {
-            values.add(value);
-            return const Placeholder();
-          },
-        ),
-      );
-
-      expect(values, isNotEmpty);
-      expect(values.first, 0.0);
-
-      await tester.pump(const Duration(milliseconds: 50));
-      expect(values.last, greaterThan(0.0));
-
-      await tester.pump(const Duration(milliseconds: 49));
-      expect(values.last, greaterThan(0.9));
-
-      await tester.pump(const Duration(milliseconds: 10));
-      expect(values.last, lessThan(0.2));
-    });
-  });
 }
